third_party/WebKit/LayoutTests/inspector-protocol/heap-profiler/heap-objects-tracking.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function junkGenerator()
{
    var junkArray = new Array(1000);
    for (var i = 0; i < junkArray.length; ++i)
        junkArray[i] = "42 " + i;
    window.junkArray = junkArray;
}

function setupIntervalAndRunTest()
{
    setInterval(junkGenerator, 0);
    runTest();
}

function test()
{
    var lastSeenObjectIdEventCount = 0;
    var heapStatsUpdateEventCount = 0;
    function trackingStarted()
    {
        InspectorTest.log("SUCCESS: tracking started");
    }

    function trackingStopped()
    {
        InspectorTest.log("Number of heapStatsUpdate events >= numbrt of lastSeenObjectId events: " + (heapStatsUpdateEventCount >= lastSeenObjectIdEventCount));
        InspectorTest.log("At least 2 lastSeenObjectId arrived: " + (lastSeenObjectIdEventCount >= 2));
        InspectorTest.log("SUCCESS: tracking stopped");
        InspectorTest.completeTest();
    }

    var fragments = [];
    InspectorTest.eventHandler["HeapProfiler.lastSeenObjectId"] = function(messageObject)
    {
        if (++lastSeenObjectIdEventCount <= 2) {
            var params = messageObject["params"];
            InspectorTest.log("HeapProfiler.lastSeenObjectId has params: " + !!params);
            InspectorTest.log("HeapProfiler.lastSeenObjectId has params.lastSeenObjectId: " + !!params.lastSeenObjectId);
            InspectorTest.log("HeapProfiler.lastSeenObjectId has timestamp: " + !!params.timestamp);
            InspectorTest.log("A heap stats fragment did arrive before HeapProfiler.lastSeenObjectId: " + !!fragments.length);
            InspectorTest.log("");
        }
        if (lastSeenObjectIdEventCount == 2) {
            // Wait for two updates and then stop tracing.
             InspectorTest.sendCommand("HeapProfiler.stopTrackingHeapObjects", {}, trackingStopped);
        }
    }

    InspectorTest.eventHandler["HeapProfiler.heapStatsUpdate"] = function(messageObject)
    {
        ++heapStatsUpdateEventCount
        var params = messageObject["params"];
        if (heapStatsUpdateEventCount <= 2)
            InspectorTest.log("HeapProfiler.heapStatsUpdate has params: " + !!params);
        var statsUpdate = params.statsUpdate;
        if (heapStatsUpdateEventCount <= 2) {
            InspectorTest.log("HeapProfiler.heapStatsUpdate has statsUpdate: " + !!statsUpdate);
            InspectorTest.log("statsUpdate length is not zero: " + !!statsUpdate.length);
            InspectorTest.log("statsUpdate length is a multiple of three: " + !(statsUpdate.length % 3));
            InspectorTest.log("statsUpdate: first fragmentIndex in first update: " + statsUpdate[0]);
            InspectorTest.log("statsUpdate: total count of objects is not zero: " + !!statsUpdate[1]);
            InspectorTest.log("statsUpdate: total size of objects is not zero: " + !!statsUpdate[2]);
            InspectorTest.log("");
        }
        fragments.push(statsUpdate);
    }

    InspectorTest.sendCommand("HeapProfiler.startTrackingHeapObjects", {}, trackingStarted);
    //# sourceURL=heap-objects-tracking.html
}
</script>
</head>
<body onload="setupIntervalAndRunTest()">
<p>Test that heap tracking actually reports data fragments.</p>
</body>
</html>
