third_party/blink/web_tests/inspector-protocol/heap-profiler/heap-objects-tracking-expected.txt

Test that heap tracking actually reports data fragments.
SUCCESS: tracking started
HeapProfiler.heapStatsUpdate has params: true
HeapProfiler.heapStatsUpdate has statsUpdate: true
statsUpdate length is not zero: true
statsUpdate length is a multiple of three: true
statsUpdate: first fragmentIndex in first update: 0
statsUpdate: total count of objects is not zero: true
statsUpdate: total size of objects is not zero: true

HeapProfiler.lastSeenObjectId has params: true
HeapProfiler.lastSeenObjectId has params.lastSeenObjectId: true
HeapProfiler.lastSeenObjectId has timestamp: true
A heap stats fragment did arrive before HeapProfiler.lastSeenObjectId: true

HeapProfiler.heapStatsUpdate has params: true
HeapProfiler.heapStatsUpdate has statsUpdate: true
statsUpdate length is not zero: true
statsUpdate length is a multiple of three: true
statsUpdate: first fragmentIndex in first update: 0
statsUpdate: total count of objects is not zero: true
statsUpdate: total size of objects is not zero: true

HeapProfiler.lastSeenObjectId has params: true
HeapProfiler.lastSeenObjectId has params.lastSeenObjectId: true
HeapProfiler.lastSeenObjectId has timestamp: true
A heap stats fragment did arrive before HeapProfiler.lastSeenObjectId: true

Number of heapStatsUpdate events >= number of lastSeenObjectId events: true
At least 2 lastSeenObjectId arrived: true
SUCCESS: tracking stopped
